backend/requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langchain-huggingface
sentence-transformers
beautifulsoup4
pypdf
firebase-admin
python-multipart
python-dotenv
langchain-community
qdrant-client
langchain-qdrant
